cybersecurity/agents/network_monitor/requirements.txt
confluent-kafka
elasticsearch
redis
pydantic
loguru
python-dotenv
scapy
pyshark
